Ultimazione classe Storia e modifiche css
Ultimazione classe Storia, modifiche css, aggiunta controlli sull'inserimento del file

+++ Instacam/gestoreAddStoria.php
@@ -1,12 +1,20 @@
 <?php
     require_once "Classi/Storia.php";
 
-    if (isset($_FILES['file'])) 
+    if (isset($_FILES["file"]) && !empty($_FILES["file"])) 
     {
+        
 
         //prendo info dal file
-        $fileTmpPath = $_FILES['file']['tmp_name'];
-        $fileName = $_FILES['file']['name'];
+        $fileTmpPath = $_FILES["file"]["tmp_name"];
+        $fileName = $_FILES["file"]["name"];
+
+        if($fileTmpPath == null)
+        {
+            header("location: paginaAddStoria.php?messaggio=nessun file caricato!");
+            exit;
+        }
+            
       
         $estensioneFile = explode(".", $fileName)[1];
         //
@@ -30,10 +38,17 @@
         //
 
         //sposta il file e stampa il risultato
-        if (move_uploaded_file($fileTmpPath, $pathFinale))
-            echo "Successo";
-        else
-            echo "Errore";
+        if (file_exists($fileTmpPath) && is_file($fileTmpPath) && is_readable($fileTmpPath))
+        {
+            if (move_uploaded_file($fileTmpPath, $pathFinale))
+                echo "Successo";
+            else
+                echo "Errore";
+        }
+        else 
+        {
+            echo "Il path non esiste.\n";
+        }
         //
 
         //prendo la data di pubblicazione e il tempo corrente
@@ -41,9 +56,15 @@
         //
 
         //aggiungo una nuova dupla nel file FilePubblicazioni.csv
-        $storia  = new Storia($estensioneFile, $infoDataPubblicazione, $username);
+        $storia  = new Storia($username, $estensioneFile, $infoDataPubblicazione);
         $storia->toCSV();
+        header("location: paginaHome.php?popup=storia caricata con successo");
     } 
     else 
+    {
         echo "No file caricati";
+        header("location: paginaLogin.php?messaggio=nessun file caricato!");
+    }
+    exit;
+    
 ?>
